// src/enums/user.ts
/* eslint-disable no-unused-vars */
export enum ENUM_USER_ROLE {
  SUPER_ADMIN = 'super_admin',
  ADMIN = 'admin',
  ACCOUNT_ADMIN = 'account_admin',
  SALESMAN = 'salesman',
  STORE_INCHARGE = 'store_incharge',
  GENERAL_USER = 'general_user',
}

export enum ENUM_USER_PERMISSION {
  CREATE_SELL = 'create-sell',
  CREATE_PRODUCT = 'create-product',
  CREATE_PURCHASE = 'create-purchase',
  REFUND_SELL = 'refund-sell',
  REFUND_PURCHASE = 'refund-purchase'
}